+++ extensions/selector/settings.ts
@@ -1,0 +1,3 @@
+import { BeatPerMinute } from '../../beatmap/shared/bpm.ts';
+
+export let BPM: BeatPerMinute | undefined | null;